requirements.txt
streamlit
python-dotenv
langchain-google-genai
langchain-core
pyttsx3
SpeechRecognition
pypdf
streamlit-mic-recorder
pytesseract
Pillow
langchain-text-splitters
